+++ interpret/hex.go
@@ -1,0 +1,113 @@
+package interpret
+
+import (
+	"encoding/hex"
+	"fmt"
+	"strconv"
+	"strings"
+)
+
+//-------------------------------------------------------
+// hex and ints
+
+func FromHex(s string) []byte {
+	if len(s) > 1 {
+		if s[0:2] == "0x" {
+			s = s[2:]
+		}
+		if len(s)%2 == 1 {
+			s = "0" + s
+		}
+		return Hex2Bytes(s)
+	}
+	return nil
+}
+
+func Hex2Bytes(str string) []byte {
+	h, _ := hex.DecodeString(str)
+
+	return h
+}
+
+func Bytes2Hex(d []byte) string {
+	return hex.EncodeToString(d)
+}
+
+// s can be string, hex, or int.
+// returns properly formatted 32byte hex value
+func Coerce2Hex(s string) string {
+	//fmt.Println("coercing to hex:", s)
+	// is int?
+	i, err := strconv.Atoi(s)
+	if err == nil {
+		return "0x" + hex.EncodeToString(NumberToBytes(int32(i), i/256+1))
+	}
+	// is already prefixed hex?
+	if len(s) > 1 && s[:2] == "0x" {
+		if len(s)%2 == 0 {
+			return s
+		}
+		return "0x0" + s[2:]
+	}
+	// is unprefixed hex?
+	if len(s) > 32 {
+		return "0x" + s
+	}
+	pad := strings.Repeat("\x00", (32-len(s))) + s
+	ret := "0x" + hex.EncodeToString([]byte(pad))
+	//fmt.Println("result:", ret)
+	return ret
+}
+
+func CoerceHexAndPad(aa string, padright bool) string {
+	if !IsHex(aa) {
+		//first try and convert to int
+		n, err := strconv.Atoi(aa)
+		if err != nil {
+			// right pad strings
+			if padright {
+				aa = "0x" + fmt.Sprintf("%x", aa) + fmt.Sprintf("%0"+strconv.Itoa(64-len(aa)*2)+"s", "")
+			} else {
+				aa = "0x" + fmt.Sprintf("%x", aa)
+			}
+		} else {
+			aa = "0x" + fmt.Sprintf("%x", n)
+		}
+	}
+	return aa
+}
+
+func IsHex(s string) bool {
+	if len(s) < 2 {
+		return false
+	}
+	if s[:2] == "0x" {
+		return true
+	}
+	return false
+}
+
+func AddHex(s string) string {
+	if len(s) < 2 {
+		return "0x" + s
+	}
+
+	if s[:2] != "0x" {
+		return "0x" + s
+	}
+
+	return s
+}
+
+func StripHex(s string) string {
+	if len(s) > 1 {
+		if s[:2] == "0x" {
+			s = s[2:]
+			if len(s)%2 != 0 {
+				s = "0" + s
+			}
+			return s
+		}
+	}
+	return s
+}
