+++ interpret/strings.go
@@ -1,0 +1,42 @@
+package interpret
+
+func LeftPadString(str string, l int) string {
+	if l < len(str) {
+		return str
+	}
+
+	zeros := Bytes2Hex(make([]byte, (l-len(str))/2))
+
+	return zeros + str
+
+}
+
+func RightPadString(str string, l int) string {
+	if l < len(str) {
+		return str
+	}
+
+	zeros := Bytes2Hex(make([]byte, (l-len(str))/2))
+
+	return str + zeros
+}
+
+func StripZeros(s string) string {
+	i := 0
+	for ; i < len(s); i++ {
+		if s[i] != '0' {
+			break
+		}
+	}
+	return s[i:]
+}
+
+func StripOnes(s string) string {
+	i := 0
+	for ; i < len(s); i++ {
+		if s[i:i+1] != "01" {
+			break
+		}
+	}
+	return s[i:]
+}
