--- pkgs/abi/util.go
@@ -1,102 +1,0 @@
-package abi
-
-import (
-	"encoding/hex"
-	"math/big"
-)
-
-const (
-	HashLength    = 32
-	AddressLength = 20
-)
-
-type (
-	Hash    [HashLength]byte
-	Address [AddressLength]byte
-)
-
-func BytesToBig(data []byte) *big.Int {
-	n := new(big.Int)
-	n.SetBytes(data)
-
-	return n
-}
-
-func BytesToHash(b []byte) Hash {
-	var h Hash
-	h.SetBytes(b)
-	return h
-}
-
-// Sets the hash to the value of b. If b is larger than len(h) it will panic
-func (h *Hash) SetBytes(b []byte) {
-	if len(b) > len(h) {
-		b = b[len(b)-HashLength:]
-	}
-
-	copy(h[HashLength-len(b):], b)
-}
-
-/////////// Address
-func BytesToAddress(b []byte) Address {
-	var a Address
-	a.SetBytes(b)
-	return a
-}
-
-// Sets the address to the value of b. If b is larger than len(a) it will panic
-func (a *Address) SetBytes(b []byte) {
-	if len(b) > len(a) {
-		b = b[len(b)-AddressLength:]
-	}
-	copy(a[AddressLength-len(b):], b)
-}
-
-func Hex2Bytes(str string) []byte {
-	h, _ := hex.DecodeString(str)
-
-	return h
-}
-
-func HexToAddress(s string) Address { return BytesToAddress(FromHex(s)) }
-
-func FromHex(s string) []byte {
-	if len(s) > 1 {
-		if s[0:2] == "0x" {
-			s = s[2:]
-		}
-		if len(s)%2 == 1 {
-			s = "0" + s
-		}
-		return Hex2Bytes(s)
-	}
-	return nil
-}
-
-func Bytes2Hex(d []byte) string {
-	return hex.EncodeToString(d)
-}
-
-func LeftPadString(str string, l int) string {
-	if l < len(str) {
-		return str
-	}
-
-	zeros := Bytes2Hex(make([]byte, (l-len(str))/2))
-
-	return zeros + str
-
-}
-
-func RightPadString(str string, l int) string {
-	if l < len(str) {
-		return str
-	}
-
-	zeros := Bytes2Hex(make([]byte, (l-len(str))/2))
-
-	return str + zeros
-}
-
-func (a Address) Str() string   { return string(a[:]) }
-func (a Address) Bytes() []byte { return a[:] }
