+++ util/readers.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"path"
 
-	"github.com/eris-ltd/eris-cli/hex"
+	"github.com/eris-ltd/eris-cli/interpret"
 	"github.com/eris-ltd/eris-cli/log"
 
 	"github.com/eris-ltd/eris-db/client/core"
@@ -50,7 +50,7 @@ func ReadTxSignAndBroadcast(result *core.TxResult, err error) error {
 }
 
 func ReadAbi(root, contract string) (string, error) {
-	p := path.Join(root, hex.StripHex(contract))
+	p := path.Join(root, interpret.StripHex(contract))
 	if _, err := os.Stat(p); err != nil {
 		return "", fmt.Errorf("Abi doesn't exist for =>\t%s", p)
 	}
